refactor: set auth and register controller in common folder
