+++ client/src/about.ts
@@ -1,6 +1,6 @@
 import { AppEvents } from "./components/controller/app-events";
-import AuthController from "./components/controller/index/auth-controller";
-import RegisterController from "./components/controller/index/register-controller";
+import AuthController from "./components/controller/common/auth-controller";
+import RegisterController from "./components/controller/common/register-controller";
 import Observer from "./components/controller/observer";
 import UserModel from "./components/models/index/data-model/user-model";
 import AuthModel from "./components/models/index/state-model/auth-model";
