+++ client/src/components/controller/common/register-controller.ts
